step-12 add-edit with validations

+++ src/main/java/org/academiadecodigo/javabank/services/CustomerServiceImpl.java
@@ -1,11 +1,16 @@
 package org.academiadecodigo.javabank.services;
 
-import org.academiadecodigo.javabank.persistence.model.Customer;
-import org.academiadecodigo.javabank.persistence.model.Recipient;
-import org.academiadecodigo.javabank.persistence.model.account.Account;
+import org.academiadecodigo.javabank.exceptions.AccountNotFoundException;
+import org.academiadecodigo.javabank.exceptions.AssociationExistsException;
+import org.academiadecodigo.javabank.exceptions.CustomerNotFoundException;
+import org.academiadecodigo.javabank.exceptions.RecipientNotFoundException;
 import org.academiadecodigo.javabank.persistence.dao.AccountDao;
 import org.academiadecodigo.javabank.persistence.dao.CustomerDao;
 import org.academiadecodigo.javabank.persistence.dao.RecipientDao;
+import org.academiadecodigo.javabank.persistence.model.AbstractModel;
+import org.academiadecodigo.javabank.persistence.model.Customer;
+import org.academiadecodigo.javabank.persistence.model.Recipient;
+import org.academiadecodigo.javabank.persistence.model.account.Account;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,8 +18,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-
-import static org.academiadecodigo.javabank.errors.ErrorMessage.*;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * An {@link CustomerService} implementation
@@ -59,19 +64,26 @@ public class CustomerServiceImpl implements CustomerService {
     /**
      * @see CustomerService#get(Integer)
      */
-    public Customer get(Integer id) {
-        return customerDao.findById(id);
+    @Override
+    public Customer get(Integer id) throws CustomerNotFoundException {
+
+        Customer customer = customerDao.findById(id);
+
+        if (customer == null) {
+            throw new CustomerNotFoundException();
+        }
+
+        return customer;
     }
 
     /**
      * @see CustomerService#getBalance(Integer)
      */
     @Override
-    public double getBalance(Integer id) {
+    public double getBalance(Integer id) throws CustomerNotFoundException {
 
         Customer customer = Optional.ofNullable(customerDao.findById(id))
-                .orElseThrow(() -> new IllegalArgumentException(CUSTOMER_NOT_FOUND));
-
+                .orElseThrow(CustomerNotFoundException::new);
 
         return customer.getAccounts().stream()
                 .mapToDouble(Account::getBalance)
@@ -79,11 +91,28 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     /**
+     * @see CustomerService#save(Customer)
+     */
+    @Transactional
+    @Override
+    public Customer save(Customer customer) {
+        return customerDao.saveOrUpdate(customer);
+    }
+
+    /**
      * @see CustomerService#delete(Integer)
      */
     @Transactional
     @Override
-    public void delete(Integer id) {
+    public void delete(Integer id) throws CustomerNotFoundException, AssociationExistsException {
+
+        Customer customer = Optional.ofNullable(customerDao.findById(id))
+                .orElseThrow(CustomerNotFoundException::new);
+
+        if (!customer.getAccounts().isEmpty()) {
+            throw new AssociationExistsException();
+        }
+
         customerDao.delete(id);
     }
 
@@ -100,14 +129,35 @@ public class CustomerServiceImpl implements CustomerService {
      */
     @Transactional(readOnly = true)
     @Override
-    public List<Recipient> listRecipients(Integer id) {
+    public List<Recipient> listRecipients(Integer id) throws CustomerNotFoundException {
 
         // check then act logic requires transaction,
         // event if read only
+
         Customer customer = Optional.ofNullable(customerDao.findById(id))
-                .orElseThrow(() -> new IllegalArgumentException(CUSTOMER_NOT_FOUND));
+                .orElseThrow(CustomerNotFoundException::new);
 
         return new ArrayList<>(customer.getRecipients());
+    }
+
+    /**
+     * @see CustomerService#addRecipient(Integer, Recipient)
+     */
+    @Transactional
+    @Override
+    public void addRecipient(Integer id, Recipient recipient) throws CustomerNotFoundException, AccountNotFoundException {
+
+        Customer customer = Optional.ofNullable(customerDao.findById(id))
+                .orElseThrow(CustomerNotFoundException::new);
+
+        if (accountDao.findById(recipient.getAccountNumber()) == null ||
+                getAccountIds(customer).contains(recipient.getAccountNumber())) {
+            throw new AccountNotFoundException();
+        }
+
+        customer.addRecipient(recipient);
+        customerDao.saveOrUpdate(customer);
+
     }
 
     /**
@@ -115,19 +165,27 @@ public class CustomerServiceImpl implements CustomerService {
      */
     @Transactional
     @Override
-    public void removeRecipient(Integer id, Integer recipientId) {
+    public void removeRecipient(Integer id, Integer recipientId) throws CustomerNotFoundException, RecipientNotFoundException {
 
         Customer customer = Optional.ofNullable(customerDao.findById(id))
-                .orElseThrow(() -> new IllegalArgumentException(CUSTOMER_NOT_FOUND));
+                .orElseThrow(CustomerNotFoundException::new);
 
         Recipient recipient = Optional.ofNullable(recipientDao.findById(recipientId))
-                .orElseThrow(() -> new IllegalArgumentException(RECIPIENT_NOT_FOUND));
+                .orElseThrow(RecipientNotFoundException::new);
 
-        if (!recipient.getCustomer().getId().equals(id)) {
-            throw new IllegalArgumentException(CUSTOMER_RECIPIENT_NOT_FOUND);
+        if (!customer.getRecipients().contains(recipient)) {
+            throw new RecipientNotFoundException();
         }
 
         customer.removeRecipient(recipient);
         customerDao.saveOrUpdate(customer);
+    }
+
+    private Set<Integer> getAccountIds(Customer customer) {
+        List<Account> accounts = customer.getAccounts();
+
+        return accounts.stream()
+                .map(AbstractModel::getId)
+                .collect(Collectors.toSet());
     }
 }
